Split public API tests from internal ones
This makes it more difficult to forget to mark public API with `pub`.
